feat: move logger types to a separate package
