+++ packages/core/src/plugins/router/modules/response.ts
@@ -15,7 +15,8 @@ import {
   TimeoutError,
   ValidationError
 } from 'common-errors'
-import { Microfleet, LoggerPlugin } from '../../../'
+import { Microfleet } from '../../../'
+import { LoggerPlugin } from '@microfleet/types/types/plugin-logger'
 import { ServiceRequest } from '../../../types'
 import moduleLifecycle from './lifecycle'
