+++ packages/core/tsconfig.build.json
@@ -4,6 +4,9 @@
     "rootDir": "./src",
     "outDir": "./lib"
   },
+  "references": [{
+    "path": "../types/tsconfig.build.json"
+  }],
   "include": [
     "./src/*.ts",
     "./src/**/*.ts"
